requirements.txt
matplotlib==3.7.2
numpy==1.25.1
pandas==2.0.3
python-dotenv==1.0.0
scikit_learn==1.3.0
seaborn==0.12.2
tqdm==4.65.0
